pyproject.toml
[project]
name = "enterprise-context-platform"
version = "3.0.0"
description = "Cross-estate semantic mediation layer for enterprise agentic AI"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "asyncpg>=0.29.0",
    "neo4j>=5.17.0",
    "redis>=5.0.0",
    "httpx>=0.27.0",
    "anthropic>=0.47.0",      # Claude — intent parsing in intelligent mode
    "openai>=1.12.0",         # OpenAI — embeddings only (text-embedding-3-small)
    "numpy>=1.26.0",
    "pgvector>=0.2.5",        # Precedent + asset embedding search
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
